refactor(lsp): split signature-help into trigger/candidate/overload/display phases (#9418)
